Move module `matrixtest` to `matrix.test`
