+++ app/build.gradle.kts
@@ -221,5 +221,5 @@ dependencies {
     testImplementation(libs.molecule.runtime)
     testImplementation(libs.test.truth)
     testImplementation(libs.test.turbine)
-    testImplementation(projects.libraries.matrixtest)
+    testImplementation(projects.libraries.matrix.test)
 }
