+++ features/login/build.gradle.kts
@@ -48,7 +48,7 @@ dependencies {
     testImplementation(libs.molecule.runtime)
     testImplementation(libs.test.truth)
     testImplementation(libs.test.turbine)
-    testImplementation(projects.libraries.matrixtest)
+    testImplementation(projects.libraries.matrix.test)
 
     androidTestImplementation(libs.test.junitext)
 }
