+++ features/messages/src/test/kotlin/io/element/android/features/messages/actionlist/ActionListPresenterTest.kt
@@ -29,15 +29,14 @@ import io.element.android.features.messages.timeline.model.event.TimelineItemEve
 import io.element.android.features.messages.timeline.model.event.TimelineItemRedactedContent
 import io.element.android.features.messages.timeline.model.event.TimelineItemTextContent
 import io.element.android.libraries.designsystem.components.avatar.AvatarData
-import io.element.android.libraries.matrixtest.AN_EVENT_ID
-import io.element.android.libraries.matrixtest.A_MESSAGE
-import io.element.android.libraries.matrixtest.A_USER_ID
-import io.element.android.libraries.matrixtest.A_USER_NAME
+import io.element.android.libraries.matrix.test.AN_EVENT_ID
+import io.element.android.libraries.matrix.test.A_MESSAGE
+import io.element.android.libraries.matrix.test.A_USER_ID
+import io.element.android.libraries.matrix.test.A_USER_NAME
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import org.junit.Test
-import org.matrix.rustcomponents.sdk.TimelineItemContent
 
 class ActionListPresenterTest {
     @Test
