+++ features/messages/src/test/kotlin/io/element/android/features/messages/textcomposer/MessageComposerPresenterTest.kt
@@ -24,12 +24,12 @@ import app.cash.turbine.ReceiveTurbine
 import app.cash.turbine.test
 import com.google.common.truth.Truth.assertThat
 import io.element.android.libraries.core.data.StableCharSequence
-import io.element.android.libraries.matrixtest.ANOTHER_MESSAGE
-import io.element.android.libraries.matrixtest.AN_EVENT_ID
-import io.element.android.libraries.matrixtest.A_MESSAGE
-import io.element.android.libraries.matrixtest.A_REPLY
-import io.element.android.libraries.matrixtest.A_USER_NAME
-import io.element.android.libraries.matrixtest.room.FakeMatrixRoom
+import io.element.android.libraries.matrix.test.ANOTHER_MESSAGE
+import io.element.android.libraries.matrix.test.AN_EVENT_ID
+import io.element.android.libraries.matrix.test.A_MESSAGE
+import io.element.android.libraries.matrix.test.A_REPLY
+import io.element.android.libraries.matrix.test.A_USER_NAME
+import io.element.android.libraries.matrix.test.room.FakeMatrixRoom
 import io.element.android.libraries.textcomposer.MessageComposerMode
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
