+++ features/preferences/src/test/kotlin/io/element/android/features/preferences/root/PreferencesRootPresenterTest.kt
@@ -25,7 +25,7 @@ import com.google.common.truth.Truth.assertThat
 import io.element.android.features.logout.LogoutPreferencePresenter
 import io.element.android.features.rageshake.preferences.RageshakePreferencesPresenter
 import io.element.android.libraries.architecture.Async
-import io.element.android.libraries.matrixtest.FakeMatrixClient
+import io.element.android.libraries.matrix.test.FakeMatrixClient
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import org.junit.Test
