+++ features/rageshake/build.gradle.kts
@@ -51,7 +51,7 @@ dependencies {
     testImplementation(libs.molecule.runtime)
     testImplementation(libs.test.truth)
     testImplementation(libs.test.turbine)
-    testImplementation(projects.libraries.matrixtest)
+    testImplementation(projects.libraries.matrix.test)
     testImplementation(libs.test.mockk)
 
     androidTestImplementation(libs.test.junitext)
