+++ features/rageshake/src/test/kotlin/io/element/android/features/rageshake/bugreport/BugReportPresenterTest.kt
@@ -25,7 +25,7 @@ import com.google.common.truth.Truth.assertThat
 import io.element.android.features.rageshake.crash.ui.A_CRASH_DATA
 import io.element.android.features.rageshake.crash.ui.FakeCrashDataStore
 import io.element.android.libraries.architecture.Async
-import io.element.android.libraries.matrixtest.A_FAILURE_REASON
+import io.element.android.libraries.matrix.test.A_FAILURE_REASON
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import org.junit.Test
