+++ features/roomlist/build.gradle.kts
@@ -49,7 +49,7 @@ dependencies {
     testImplementation(libs.molecule.runtime)
     testImplementation(libs.test.truth)
     testImplementation(libs.test.turbine)
-    testImplementation(projects.libraries.matrixtest)
+    testImplementation(projects.libraries.matrix.test)
 
     //testImplementation(testFixtures(projects.libraries.matrix))
