+++ features/roomlist/src/test/kotlin/io/element/android/features/roomlist/RoomListPresenterTests.kt
@@ -26,17 +26,16 @@ import io.element.android.features.roomlist.model.RoomListEvents
 import io.element.android.features.roomlist.model.RoomListRoomSummary
 import io.element.android.libraries.dateformatter.LastMessageFormatter
 import io.element.android.libraries.designsystem.components.avatar.AvatarData
-import io.element.android.libraries.matrixtest.AN_AVATAR_URL
-import io.element.android.libraries.matrixtest.AN_EXCEPTION
-import io.element.android.libraries.matrixtest.A_MESSAGE
-import io.element.android.libraries.matrixtest.A_ROOM_ID
-import io.element.android.libraries.matrixtest.A_ROOM_NAME
-import io.element.android.libraries.matrixtest.A_SESSION_ID
-import io.element.android.libraries.matrixtest.A_USER_ID
-import io.element.android.libraries.matrixtest.A_USER_NAME
-import io.element.android.libraries.matrixtest.FakeMatrixClient
-import io.element.android.libraries.matrixtest.room.FakeRoomSummaryDataSource
-import io.element.android.libraries.matrixtest.room.aRoomSummaryFilled
+import io.element.android.libraries.matrix.test.AN_AVATAR_URL
+import io.element.android.libraries.matrix.test.AN_EXCEPTION
+import io.element.android.libraries.matrix.test.A_MESSAGE
+import io.element.android.libraries.matrix.test.A_ROOM_ID
+import io.element.android.libraries.matrix.test.A_ROOM_NAME
+import io.element.android.libraries.matrix.test.A_USER_ID
+import io.element.android.libraries.matrix.test.A_USER_NAME
+import io.element.android.libraries.matrix.test.FakeMatrixClient
+import io.element.android.libraries.matrix.test.room.FakeRoomSummaryDataSource
+import io.element.android.libraries.matrix.test.room.aRoomSummaryFilled
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import org.junit.Test
