+++ libraries/matrix/test/src/main/kotlin/io/element/android/libraries/matrix/test/FakeMatrixClient.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.element.android.libraries.matrixtest
+package io.element.android.libraries.matrix.test
 
 import io.element.android.libraries.matrix.MatrixClient
 import io.element.android.libraries.matrix.core.RoomId
@@ -22,9 +22,9 @@ import io.element.android.libraries.matrix.core.SessionId
 import io.element.android.libraries.matrix.media.MediaResolver
 import io.element.android.libraries.matrix.room.MatrixRoom
 import io.element.android.libraries.matrix.room.RoomSummaryDataSource
-import io.element.android.libraries.matrixtest.media.FakeMediaResolver
-import io.element.android.libraries.matrixtest.room.FakeMatrixRoom
-import io.element.android.libraries.matrixtest.room.FakeRoomSummaryDataSource
+import io.element.android.libraries.matrix.test.media.FakeMediaResolver
+import io.element.android.libraries.matrix.test.room.FakeMatrixRoom
+import io.element.android.libraries.matrix.test.room.FakeRoomSummaryDataSource
 import kotlinx.coroutines.delay
 import org.matrix.rustcomponents.sdk.MediaSource
