+++ libraries/matrix/test/src/main/kotlin/io/element/android/libraries/matrix/test/timeline/FakeMatrixTimeline.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.element.android.libraries.matrixtest.timeline
+package io.element.android.libraries.matrix.test.timeline
 
 import io.element.android.libraries.matrix.core.EventId
 import io.element.android.libraries.matrix.timeline.MatrixTimeline
