+++ settings.gradle.kts
@@ -61,7 +61,6 @@ include(":tests:uitests")
 include(":anvilannotations")
 include(":anvilcodegen")
 include(":libraries:architecture")
-include(":libraries:matrixtest")
 include(":features:template")
 include(":libraries:androidutils")
 include(":samples:minimal")
